fix: [#7] missing commit signature
